+++ CHANGELOG.md
@@ -9,9 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Single pole Drude-Lorentz dispersion relation.
-* Double pole Drude-Lorentz dispersion relation.
+* Single pole Drude-Lorentz permittivity.
+* Double pole Drude-Lorentz permittivity.
 * Surface plasmon polariton dispersion relation.
+* Utilities:
+    * permittivity to extinction coefficient.
+    * permittivity to refractive index.
+    * refractive index to permittivity.
 
 ### Changed